Pickups complete & slot coordinate system created
The slot coordinate system makes it so that we can add items to an inventory and determine the item's slot position, even if the slot visuals aren't created yet. That way when we open containers or an NPC's inventory, the item's will already have determined positions visually within the inventory. This also helps us determine if an item is too large to fit into an inventory.